add smarthub dataaccess library for shared data access abstractions add edge infrastructure library
