+++ src/DataAccess/SmartHub.DataAccess/Concrete/Repository.cs
@@ -3,10 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using Microsoft.EntityFrameworkCore;
-using SmartHub.Domain.Contracts;
-using SmartHub.Infrastructure.Contracts;
+using SmartHub.DataAccess.Contracts;
 
-namespace SmartHub.Infrastructure.Concrete
+namespace SmartHub.DataAccess.Concrete
 {
   public class Repository<TEntity> : IRepository<TEntity> where TEntity : class
   {
