+++ src/DataAccess/SmartHub.DataAccess/Contracts/IUnitOfWork.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace SmartHub.DataAccess.Contracts
+{
+  public interface IUnitOfWorkBase : IDisposable
+  {
+    int Commit();
+  }
+}
