+++ src/Devices/SmartHub.BusinessLogic/Helpers/TimeConverterExtensions.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace SmartHub.BL.Helpers
 {
