--- src/Devices/SmartHub.BusinessLogic/Models/TimeSeries.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace SmartHub.BusinessLogic.Models
-{
-  public class TimeSeries
-  {
-    public IEnumerable<Measurement> Values { get; set; }
-  }
-}
