+++ src/Devices/SmartHub.Infrastructure/Contracts/IDeviceRepository.cs
@@ -1,5 +1,5 @@
 using SmartHub.Infrastructure.Entities;
-
+using SmartHub.DataAccess.Contracts;
 namespace SmartHub.Infrastructure.Contracts
 {
   public interface IDeviceRepository : IRepository<Device>
