--- src/Devices/SmartHub.Infrastructure/Entities/Measurement.cs
@@ -1,18 +1,0 @@
-﻿using System;
-
-namespace SmartHub.Infrastructure.Entities
-{
-  public class Measurement
-  {
-    public int Id { get; set; }
-
-    public int Type { get; set; }
-
-    public double Value { get; set; }
-
-    public DateTime? DtSend { get; set; }
-
-    public Guid DeviceId { get; set; }
-
-  }
-}
