+++ src/Devices/SmartHub.Infrastructure/SmartHubContext.cs
@@ -10,16 +10,12 @@ namespace SmartHub.Infrastructure
 
     public DbSet<Device> Devices { get; set; }
 
-    public DbSet<Measurement> Telemetries { get; set; }
-
-
     public SmartHubContext(DbContextOptions<SmartHubContext> options) : base(options) { }
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
       base.OnModelCreating(modelBuilder);
       modelBuilder.ApplyConfiguration(new DeviceConfiguration());
-      modelBuilder.ApplyConfiguration(new MeasurementConfiguration());
 
       modelBuilder.ToSnakeCaseNamingConventions();
     }
