+++ src/Edge/SmartHub.Edge.Infrastructure/Entities/Measurement.cs
@@ -4,6 +4,9 @@ namespace SmartHub.BusinessLogic.Models
 {
   public class Measurement
   {
+
+
+
     public DateTime DtSent { get; set; }
 
     public double Value { get; set; }
