+++ src/Edge/SmartHub.Edge.Infrastructure/Entities/MeasurementType.cs
