+++ src/Edge/SmartHub.Edge.Infrastructure/SmartHub.Edge.Infrastructure.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Configuration\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\DataAccess\SmartHub.DataAccess\SmartHub.DataAccess.csproj" />
+  </ItemGroup>
+
+</Project>
